resolution is unsigned int

+++ include/model.h
@@ -31,7 +31,7 @@ class Model{
     void setRadiusMap(std::unordered_map<std::string, double> map);
     void debug();
 	std::vector<uint8_t> getMatrix();
-	std::array<double,3> getResolution();
+	std::array<unsigned int,3> getResolution();
   private:
     std::vector<std::tuple<std::string, double, double, double>> raw_atom_coordinates;
     std::unordered_map<std::string, double> raw_radius_map;
